django_admin/requirements.txt
django==4.2.11
django-split-settings==1.3.1
python-dotenv==1.0.1
flake8==6.1.0
ruff==0.4.9
psycopg2-binary==2.9.9
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
requests==2.32.3
uwsgi==2.0.26
django-debug-toolbar==4.4.2
django-cors-headers==4.4.0
